DELIA-40553: Duplicate timedMetadata reported and wrong time value
Reason for change: During linear playlist refresh , already reported Metadata
was getting reported .Also intermittenly time of Metadata was showing outside
culling window. Fixed the duplicate handling across playlist refresh.On underflow
retune, culled value was resetted to 0 causing timed Metadata time to reset.
Test Procedure: Tested with Peacock linear content and checked Metadata
Risks: Low

(cherry picked from commit d147f3cfb18520f68d5576d22ca0823257d56193)

+++ fragmentcollector_mpd.cpp
@@ -3873,7 +3873,7 @@ void PrivateStreamAbstractionMPD::ProcessPeriodSupplementalProperty(Node* node, 
 					{
 						const std::string& tag = aamp->subscribedTags.at(i);
 						if (tag == "#EXT-X-CUE") {
-							aamp->ReportTimedMetadata(startMS, tag.c_str(), content.c_str(), content.size());
+							aamp->ReportTimedMetadata((long long)startMS, tag.c_str(), content.c_str(), content.size());
 							break;
 						}
 					}
@@ -3952,7 +3952,7 @@ void PrivateStreamAbstractionMPD::ProcessPeriodAssetIdentifier(Node* node, uint6
 					{
 						const std::string& tag = aamp->subscribedTags.at(i);
 						if (tag == "#EXT-X-ASSET-ID") {
-							aamp->ReportTimedMetadata(startMS, tag.c_str(), content.c_str(), content.size());
+							aamp->ReportTimedMetadata((long long)startMS, tag.c_str(), content.c_str(), content.size());
 							break;
 						}
 					}
@@ -4090,7 +4090,7 @@ void PrivateStreamAbstractionMPD::ProcessTrickModeRestriction(Node* node, const 
 		{
 			const std::string& tag = aamp->subscribedTags.at(i);
 			if (tag == "#EXT-X-TRICKMODE-RESTRICTION") {
-				aamp->ReportTimedMetadata(startMS, tag.c_str(), content.c_str(), content.size());
+				aamp->ReportTimedMetadata((long long)startMS, tag.c_str(), content.c_str(), content.size());
 				break;
 			}
 		}
